AI-Fitness-Coach-Tailoring-Workout-Plans-with-Large-Language-Models/requirements.txt
streamlit
langgraph
langchain-core
langchain-google-genai
python-dotenv